+++ rye/src/cli/toolchain.rs
@@ -97,12 +97,11 @@ fn register(cmd: RegisterCommand) -> Result<(), Error> {
     let target_version = match cmd.name {
         Some(ref name) => format!("{}@{}", name, info.python_version),
         None => {
-            let name = if info.python_implementation.eq_ignore_ascii_case("cpython") {
-                "custom-cpython"
-            } else {
-                &info.python_implementation
-            };
-            format!("{}@{}", name.to_ascii_lowercase(), info.python_version)
+            format!(
+                "{}@{}",
+                info.python_implementation.to_ascii_lowercase(),
+                info.python_version
+            )
         }
     };
     let target_version: PythonVersion = target_version.parse()?;
@@ -117,10 +116,29 @@ fn register(cmd: RegisterCommand) -> Result<(), Error> {
         fs::create_dir_all(parent).ok();
     }
 
-    // XXX: this requires elevated privileges on windows but using a hardlink here would
-    // break the experience because then the interpreter does not know where it's from.
-    // maybe we want to place files there containing the path to the interpreter instead.
-    symlink_file(&cmd.path, target).context("could not symlink interpreter")?;
+    // on unix we always create a symlink
+    #[cfg(unix)]
+    {
+        symlink_file(&cmd.path, target).context("could not symlink interpreter")?;
+    }
+
+    // on windows on the other hand we try a symlink first, but if that fails we fall back
+    // to writing the interpreter into the text file.  This is also supported by the
+    // interpreter lookup (see: get_toolchain_python_bin).  This is done because symlinks
+    // require higher privileges.
+    #[cfg(windows)]
+    {
+        if symlink_file(&cmd.path, &target).is_err() {
+            fs::write(
+                &target,
+                cmd.path
+                    .as_os_str()
+                    .to_str()
+                    .ok_or_else(|| anyhow::anyhow!("non unicode path to interpreter"))?,
+            )
+            .context("could not register interpreter")?;
+        }
+    }
     println!("Registered {} as {}", cmd.path.display(), target_version);
 
     Ok(())
@@ -144,21 +162,25 @@ pub fn remove(cmd: RemoveCommand) -> Result<(), Error> {
 fn list(cmd: ListCommand) -> Result<(), Error> {
     let mut toolchains = list_known_toolchains()?
         .into_iter()
-        .map(|version| (version, true))
+        .map(|(version, path)| (version, Some(path)))
         .collect::<HashMap<_, _>>();
 
     if cmd.include_downloadable {
         for version in iter_downloadable(OS, ARCH) {
-            toolchains.entry(version).or_insert(false);
+            toolchains.entry(version).or_insert(None);
         }
     }
 
     let mut versions = toolchains.into_iter().collect::<Vec<_>>();
-    versions.sort_by_cached_key(|a| (!a.1, a.0.kind.to_string(), Reverse(a.clone())));
+    versions.sort_by_cached_key(|a| (a.1.is_none(), a.0.kind.to_string(), Reverse(a.clone())));
 
-    for (version, installed) in versions {
-        if installed {
-            println!("{}", style(&version).green());
+    for (version, path) in versions {
+        if let Some(path) = path {
+            println!(
+                "{} ({})",
+                style(&version).green(),
+                style(path.display()).dim()
+            );
         } else {
             println!("{} (downloadable)", style(version).dim());
         }